feat: New fromPackageJson builder
