+++ src/lib/config/project-paths.spec.ts
@@ -1,0 +1,27 @@
+import { describe, expect, it } from 'vitest';
+import * as path from 'path';
+import { cwd } from 'process';
+import { findRootTsConfigJsonCore } from './project-paths.js';
+import { createMemoryIo } from '../utils/io/__test-helpers__/memory-io.js';
+
+const ROOT = cwd();
+
+describe('findRootTsConfigJsonCore', () => {
+  const withPackageJson = () => createMemoryIo().setFile(path.join(ROOT, 'package.json'), '{}');
+
+  it('prefers tsconfig.base.json when both exist', () => {
+    const io = withPackageJson()
+      .setFile(path.join(ROOT, 'tsconfig.base.json'), '{}')
+      .setFile(path.join(ROOT, 'tsconfig.json'), '{}');
+    expect(findRootTsConfigJsonCore(io)).toBe(path.join(ROOT, 'tsconfig.base.json'));
+  });
+
+  it('falls back to tsconfig.json when no base config exists', () => {
+    const io = withPackageJson().setFile(path.join(ROOT, 'tsconfig.json'), '{}');
+    expect(findRootTsConfigJsonCore(io)).toBe(path.join(ROOT, 'tsconfig.json'));
+  });
+
+  it('throws when neither config is present', () => {
+    expect(() => findRootTsConfigJsonCore(withPackageJson())).toThrow(/Neither a tsconfig/);
+  });
+});
