+++ src/lib/config/share-utils.spec.ts
@@ -1,122 +1,10 @@
-import { describe, expect, it } from 'vitest';
+import { describe, expect, it, vi } from 'vitest';
 import * as path from 'path';
-import { cwd } from 'process';
-import { findRootTsConfigJsonCore, getSecondaries, shareAllCore, shareCore } from './share-utils.js';
-import { DEFAULT_SKIP_LIST, prepareSkipList } from './default-skip-list.js';
+import { shareAllCore, shareCore } from './share-utils.js';
+import { DEFAULT_SKIP_LIST } from './default-skip-list.js';
+import { logger } from '../utils/logger.js';
 import { createMemoryIo } from '../utils/io/__test-helpers__/memory-io.js';
 import { createPackageJsonRepository } from '../utils/io/package-json-repository.js';
-import type { ExternalConfig } from '../domain/config/external-config.contract.js';
-
-const ROOT = cwd();
-const EMPTY_SKIP = prepareSkipList([]);
-const shareObject: ExternalConfig = { singleton: true, requiredVersion: '^1.0.0' };
-
-describe('findRootTsConfigJsonCore', () => {
-  const withPackageJson = () => createMemoryIo().setFile(path.join(ROOT, 'package.json'), '{}');
-
-  it('prefers tsconfig.base.json when both exist', () => {
-    const io = withPackageJson()
-      .setFile(path.join(ROOT, 'tsconfig.base.json'), '{}')
-      .setFile(path.join(ROOT, 'tsconfig.json'), '{}');
-    expect(findRootTsConfigJsonCore(io)).toBe(path.join(ROOT, 'tsconfig.base.json'));
-  });
-
-  it('falls back to tsconfig.json when no base config exists', () => {
-    const io = withPackageJson().setFile(path.join(ROOT, 'tsconfig.json'), '{}');
-    expect(findRootTsConfigJsonCore(io)).toBe(path.join(ROOT, 'tsconfig.json'));
-  });
-
-  it('throws when neither config is present', () => {
-    expect(() => findRootTsConfigJsonCore(withPackageJson())).toThrow(/Neither a tsconfig/);
-  });
-});
-
-describe('getSecondaries', () => {
-  const LIB = path.resolve('/root/node_modules/mylib');
-
-  it('returns an empty map when the lib path does not exist', () => {
-    expect(getSecondaries(createMemoryIo(), true, LIB, 'mylib', shareObject, EMPTY_SKIP)).toEqual(
-      {}
-    );
-  });
-
-  it('reads configured secondaries from the package.json exports map', () => {
-    const io = createMemoryIo().setFile(
-      path.join(LIB, 'package.json'),
-      JSON.stringify({
-        version: '1.2.3',
-        type: 'module',
-        exports: {
-          '.': './index.js',
-          './sub': { default: './sub/index.js' },
-        },
-      })
-    );
-
-    const result = getSecondaries(io, true, LIB, 'mylib', shareObject, EMPTY_SKIP);
-
-    expect(Object.keys(result!)).toEqual(['mylib/sub']);
-    expect(result!['mylib/sub']).toMatchObject({ singleton: true });
-  });
-
-  it('resolves a glob whose pattern has no JS extension and filters non-JS files', () => {
-    const io = createMemoryIo()
-      .setFile(
-        path.join(LIB, 'package.json'),
-        JSON.stringify({
-          version: '1.2.3',
-          type: 'module',
-          exports: {
-            '.': './index.js',
-            './components/*': './components/*',
-          },
-        })
-      )
-      .setFile(path.join(LIB, 'components', 'button.js'), '')
-      .setFile(path.join(LIB, 'components', 'icon.mjs'), '')
-      .setFile(path.join(LIB, 'components', 'styles.css'), '')
-      .setFile(path.join(LIB, 'components', 'logo.svg'), '');
-
-    const result = getSecondaries(
-      io,
-      { skip: [], resolveGlob: true },
-      LIB,
-      'mylib',
-      shareObject,
-      EMPTY_SKIP
-    );
-
-    expect(Object.keys(result!).sort()).toEqual([
-      'mylib/components/button.js',
-      'mylib/components/icon.mjs',
-    ]);
-  });
-
-  it('falls back to walking subfolders (readDir) when there is no exports map', () => {
-    const io = createMemoryIo()
-      .setFile(path.join(LIB, 'sub', 'package.json'), '{}')
-      .setFile(path.join(LIB, 'node_modules', 'dep', 'package.json'), '{}');
-
-    const result = getSecondaries(io, true, LIB, 'mylib', shareObject, EMPTY_SKIP);
-
-    expect(Object.keys(result!)).toEqual(['mylib/sub']);
-  });
-
-  it('honours a custom skip list during the folder walk', () => {
-    const io = createMemoryIo().setFile(path.join(LIB, 'sub', 'package.json'), '{}');
-
-    const result = getSecondaries(
-      io,
-      { skip: ['mylib/sub'] },
-      LIB,
-      'mylib',
-      shareObject,
-      EMPTY_SKIP
-    );
-
-    expect(result).toEqual({});
-  });
-});
 
 describe('shareCore (end-to-end via injected repository)', () => {
   const PROJECT = path.resolve('/ws/app');
@@ -203,7 +91,9 @@ describe('shareAllCore (end-to-end via injected repository)', () => {
       {
         projectPath: PROJECT,
         skipList: ['skipme'],
-        overrides: { overridden: { singleton: false, requiredVersion: '~3.1.0', includeSecondaries: false } },
+        overrides: {
+          overridden: { singleton: false, requiredVersion: '~3.1.0', includeSecondaries: false },
+        },
       },
       repo
     );
@@ -211,5 +101,73 @@ describe('shareAllCore (end-to-end via injected repository)', () => {
     expect(Object.keys(result!).sort()).toEqual(['mylib', 'overridden']);
     expect(result!['mylib']).toMatchObject({ singleton: true, requiredVersion: '^1.2.3' });
     expect(result!['overridden']).toMatchObject({ singleton: false, requiredVersion: '~3.1.0' });
+  });
+
+  it('patches a shared external in place', () => {
+    const io = createMemoryIo().setFile(
+      path.join(PROJECT, 'package.json'),
+      JSON.stringify({ dependencies: { mylib: '^1.2.3' } })
+    );
+    const repo = createPackageJsonRepository(io);
+
+    const result = shareAllCore(
+      io,
+      { singleton: true, includeSecondaries: false },
+      { projectPath: PROJECT, patchList: { mylib: { singleton: false, strictVersion: true } } },
+      repo
+    );
+
+    expect(result!['mylib']).toMatchObject({
+      singleton: false,
+      strictVersion: true,
+      requiredVersion: '^1.2.3',
+    });
+  });
+
+  it('ignores a patch for an external that is not shared and warns', () => {
+    const io = createMemoryIo().setFile(
+      path.join(PROJECT, 'package.json'),
+      JSON.stringify({ dependencies: { mylib: '^1.2.3' } })
+    );
+    const repo = createPackageJsonRepository(io);
+    const warn = vi.spyOn(logger, 'warn').mockImplementation(() => {});
+
+    const result = shareAllCore(
+      io,
+      { singleton: true, includeSecondaries: false },
+      { projectPath: PROJECT, patchList: { doesnotexist: { singleton: false } } },
+      repo
+    );
+
+    expect(Object.keys(result!)).toEqual(['mylib']);
+    expect(warn).toHaveBeenCalledWith(expect.stringContaining('not a shared external'));
+    warn.mockRestore();
+  });
+
+  it('ignores a patch for an external that is shadowed by overrides and warns', () => {
+    const io = createMemoryIo().setFile(
+      path.join(PROJECT, 'package.json'),
+      JSON.stringify({ dependencies: { overridden: '^3.0.0' } })
+    );
+    const repo = createPackageJsonRepository(io);
+    const warn = vi.spyOn(logger, 'warn').mockImplementation(() => {});
+
+    const result = shareAllCore(
+      io,
+      { singleton: true, includeSecondaries: false },
+      {
+        projectPath: PROJECT,
+        overrides: {
+          overridden: { singleton: false, requiredVersion: '~3.1.0', includeSecondaries: false },
+        },
+        patchList: { overridden: { singleton: true } },
+      },
+      repo
+    );
+
+    // The override value wins; the patch is not applied.
+    expect(result!['overridden']).toMatchObject({ singleton: false, requiredVersion: '~3.1.0' });
+    expect(warn).toHaveBeenCalledWith(expect.stringContaining('mutually exclusive'));
+    warn.mockRestore();
   });
 });
