+++ src/lib/config/share-utils.ts
@@ -1,20 +1,15 @@
 import * as path from 'path';
-import { cwd } from 'process';
 import { DEFAULT_SKIP_LIST, isInSkipList, prepareSkipList } from './default-skip-list.js';
-import { type SkipList, type PreparedSkipList } from '../domain/config/skip-list.contract.js';
+import { type SkipList } from '../domain/config/skip-list.contract.js';
 import {
   sharedPackageJsonRepository,
   findDepPackageJson,
   getVersionMaps,
-  type VersionMap,
 } from '../utils/package/package-info.js';
 import type { PackageJsonRepository } from '../domain/utils/package-json.contract.js';
-import { getConfigContext } from './configuration-context.js';
 import { logger } from '../utils/logger.js';
 import { nodeIo } from '../utils/io/node-io-adapter.js';
 import type { FileReaderPort, GlobPort } from '../domain/utils/io-port.contract.js';
-
-import { resolvePackageJsonExportsWildcardCore } from '../utils/package/resolve-wildcard-keys.js';
 import type {
   ExternalConfig,
   IncludeSecondariesOptions,
@@ -23,336 +18,60 @@ import type {
   SharedExternalsConfig,
   ShareExternalsOptions,
 } from '../domain/config/external-config.contract.js';
-import type { KeyValuePair } from '../domain/utils/keyvaluepair.contract.js';
+import { findPackageJson, inferProjectPath } from './project-paths.js';
+import { isInferVersion, lookupVersion } from './version-lookup.js';
+import { addSecondaries, getSecondaries } from './secondaries.js';
 
-let inferVersion = false;
+export const fromPackageJson = (baseCfg: ShareAllExternalsOptions, projectPath?: string) => {
+  const skipList: SkipList = [...DEFAULT_SKIP_LIST];
+  let overrides: ShareExternalsOptions = {};
+  const patchList: Record<string, Partial<ExternalConfig>> = {};
 
-export function findRootTsConfigJson(): string {
-  return findRootTsConfigJsonCore(nodeIo);
-}
-
-export function findRootTsConfigJsonCore(io: FileReaderPort): string {
-  const packageJson = findPackageJson(io, cwd());
-  const projectRoot = path.dirname(packageJson);
-  const tsConfigBaseJson = path.join(projectRoot, 'tsconfig.base.json');
-  const tsConfigJson = path.join(projectRoot, 'tsconfig.json');
-
-  if (io.exists(tsConfigBaseJson)) {
-    return tsConfigBaseJson;
-  } else if (io.exists(tsConfigJson)) {
-    return tsConfigJson;
-  }
-
-  throw new Error('Neither a tsconfig.json nor a tsconfig.base.json was found');
-}
-
-function findPackageJson(io: FileReaderPort, folder: string): string {
-  while (!io.exists(path.join(folder, 'package.json')) && path.dirname(folder) !== folder) {
-    folder = path.dirname(folder);
-  }
-
-  const filePath = path.join(folder, 'package.json');
-  if (io.exists(filePath)) {
-    return filePath;
-  }
-
-  throw new Error(
-    'no package.json found. Searched the following folder and all parents: ' + folder
-  );
-}
-
-function lookupVersion(key: string, workspaceRoot: string, repo: PackageJsonRepository): string {
-  const versionMaps = getVersionMaps(workspaceRoot, workspaceRoot, repo);
-
-  for (const versionMap of versionMaps) {
-    const version = lookupVersionInMap(key, versionMap);
-
-    if (version) {
-      return version;
-    }
-  }
-
-  throw new Error(
-    `Shared Dependency ${key} has requiredVersion:'auto'. However, this dependency is not found in your package.json`
-  );
-}
-
-function lookupVersionInMap(key: string, versions: VersionMap): string | null {
-  const parts = key.split('/');
-  if (parts.length >= 2 && parts[0]!.startsWith('@')) {
-    key = parts[0] + '/' + parts[1];
-  } else {
-    key = parts[0]!;
-  }
-
-  if (!versions[key]) {
-    return null;
-  }
-  return versions[key]!;
-}
-
-function _findSecondaries(
-  io: FileReaderPort,
-  libPath: string,
-  excludes: string[],
-  shareObject: ExternalConfig,
-  acc: SharedExternalsConfig,
-  preparedSkipList: PreparedSkipList
-): void {
-  const files = io.readDir(libPath);
-
-  const secondaries = files
-    .map(f => path.join(libPath, f))
-    .filter(f => io.isDirectory(f) && !f.endsWith('node_modules'));
-
-  for (const s of secondaries) {
-    if (io.exists(path.join(s, 'package.json'))) {
-      const secondaryLibName = s.replace(/\\/g, '/').replace(/^.*node_modules[/]/, '');
-
-      const inCustomSkipList = excludes.some(
-        e =>
-          e === secondaryLibName || (e.endsWith('*') && secondaryLibName.startsWith(e.slice(0, -1)))
-      );
-      if (inCustomSkipList) continue;
-
-      if (isInSkipList(secondaryLibName, preparedSkipList)) {
-        continue;
-      }
-
-      acc[secondaryLibName] = { ...shareObject };
-    }
-
-    _findSecondaries(io, s, excludes, shareObject, acc, preparedSkipList);
-  }
-}
-
-function findSecondaries(
-  io: FileReaderPort,
-  libPath: string,
-  excludes: string[],
-  shareObject: ExternalConfig,
-  preparedSkipList: PreparedSkipList
-): SharedExternalsConfig {
-  const acc = {} as SharedExternalsConfig;
-  _findSecondaries(io, libPath, excludes, shareObject, acc, preparedSkipList);
-  return acc;
-}
-
-export function getSecondaries(
-  io: FileReaderPort & GlobPort,
-  includeSecondaries: IncludeSecondariesOptions,
-  libPath: string,
-  key: string,
-  shareObject: ExternalConfig,
-  preparedSkipList: PreparedSkipList
-): SharedExternalsConfig | null {
-  let exclude: string[] = [];
-
-  let resolveGlob = false;
-  if (typeof includeSecondaries === 'object') {
-    if (includeSecondaries.skip) {
-      if (Array.isArray(includeSecondaries.skip)) {
-        exclude = includeSecondaries.skip;
-      } else if (typeof includeSecondaries.skip === 'string') {
-        exclude = [includeSecondaries.skip];
-      }
-    }
-
-    resolveGlob = !!includeSecondaries.resolveGlob;
-  }
-
-  // const libPath = path.join(path.dirname(packagePath), 'node_modules', key);
-
-  if (!io.exists(libPath)) {
-    return {};
-  }
-
-  const configured = readConfiguredSecondaries(
-    io,
-    key,
-    libPath,
-    exclude,
-    shareObject,
-    preparedSkipList,
-    resolveGlob
-  );
-  if (configured) {
-    return configured;
-  }
-
-  // Fallback: Search folders
-  const secondaries = findSecondaries(io, libPath, exclude, shareObject, preparedSkipList);
-  return secondaries;
-}
-
-function readConfiguredSecondaries(
-  io: FileReaderPort & GlobPort,
-  parent: string,
-  libPath: string,
-  exclude: string[],
-  shareObject: ExternalConfig,
-  preparedSkipList: PreparedSkipList,
-  resolveGlob: boolean
-): SharedExternalsConfig | null {
-  const libPackageJson = path.join(libPath, 'package.json');
-
-  if (!io.exists(libPackageJson)) {
-    return null;
-  }
-
-  const packageJson = JSON.parse(io.readText(libPackageJson));
-
-  const version = packageJson['version'] as string;
-  const esm = packageJson['type'] === 'module';
-
-  const exports = packageJson['exports'] as Record<string, Record<string, string>>;
-
-  if (!exports) {
-    return null;
-  }
-
-  const keys = Object.keys(exports).filter(
-    key =>
-      key !== '.' &&
-      key !== './package.json' &&
-      key.startsWith('./') &&
-      (exports[key]?.['default'] || exports[key]?.['import'] || typeof exports[key] === 'string')
-  );
-
-  const result = {} as SharedExternalsConfig;
-  const discoveredFiles = new Set<string>();
-
-  for (const key of keys) {
-    const secondaryName = path.join(parent, key).replace(/\\/g, '/');
-
-    const inCustomSkipList = exclude.some(
-      e => e === secondaryName || (e.endsWith('*') && secondaryName.startsWith(e.slice(0, -1)))
-    );
-    if (inCustomSkipList) continue;
-
-    if (isInSkipList(secondaryName, preparedSkipList)) {
-      continue;
-    }
-
-    const entry = getDefaultEntry(exports, key);
-
-    if (typeof entry !== 'string') {
-      console.log('No entry point found for ' + secondaryName);
-      continue;
-    }
-
-    if (!key.includes('*') && !isJsFile(entry)) {
-      continue;
-    }
-
-    const items = resolveGlobSecondaries(
-      io,
-      key,
-      libPath,
-      parent,
-      secondaryName,
-      entry,
-      { discovered: discoveredFiles, skip: exclude },
-      resolveGlob
-    );
-    items.forEach(e => discoveredFiles.add(typeof e === 'string' ? e : e.value));
-
-    for (const item of items) {
-      if (typeof item === 'object') {
-        result[item.key] = {
-          ...shareObject,
-          packageInfo: {
-            entryPoint: item.value,
-            version: shareObject.version ?? version,
-            esm,
-          },
+  const builder = {
+    skip(externals: SkipList) {
+      skipList.push(...externals);
+      return builder;
+    },
+    override(externals: ShareExternalsOptions) {
+      overrides = { ...overrides, ...externals };
+      return builder;
+    },
+    patch(externals: string[], cfg: Partial<ExternalConfig>) {
+      externals.forEach(external => {
+        patchList[external] = {
+          ...(patchList[external] ?? {}),
+          ...cfg,
         };
-      } else {
-        result[item] = {
-          ...shareObject,
-        };
-      }
-    }
-  }
-
-  return result;
-}
-
-function resolveGlobSecondaries(
-  io: GlobPort,
-  key: string,
-  libPath: string,
-  parent: string,
-  secondaryName: string,
-  entry: string,
-  excludes: { discovered: Set<string>; skip: string[] },
-  resolveGlob: boolean
-): Array<string | KeyValuePair> {
-  let items: Array<string | KeyValuePair> = [];
-  if (key.includes('*')) {
-    if (!resolveGlob) return items;
-    const expanded = resolvePackageJsonExportsWildcardCore(io, key, entry, libPath);
-    items = expanded
-      .map(e => ({
-        key: path.join(parent, e.key),
-        value: path.join(libPath, e.value),
-      }))
-      .filter(i => {
-        if (!isJsFile(i.value)) {
-          return false;
-        }
-        if (
-          excludes.skip.some(e =>
-            e.endsWith('*') ? i.key.startsWith(e.slice(0, -1)) : e === i.key
-          )
-        ) {
-          return false;
-        }
-        if (excludes.discovered.has(i.value)) {
-          return false;
-        }
-        return true;
       });
-  } else {
-    items = [secondaryName];
-  }
-  return items;
-}
+      return builder;
+    },
+    get() {
+      return shareAllCore(nodeIo, baseCfg, {
+        skipList,
+        projectPath,
+        overrides,
+        patchList,
+      });
+    },
+  };
 
-function isJsFile(file: string): boolean {
-  return file.endsWith('.js') || file.endsWith('.mjs') || file.endsWith('.cjs');
-}
+  return builder;
+};
 
-function getDefaultEntry(exports: Record<string, Record<string, string>>, key: string) {
-  let entry: string | undefined = '';
-  if (typeof exports[key] === 'string') {
-    entry = exports[key] as unknown as string;
-  }
-
-  if (!entry) {
-    entry = exports[key]?.['default'];
-    if (typeof entry === 'object') {
-      entry = entry['default'];
-    }
-  }
-
-  if (!entry) {
-    entry = exports[key]?.['import'];
-    if (typeof entry === 'object') {
-      entry = entry['import'] ?? entry['default'];
-    }
-  }
-
-  if (!entry) {
-    entry = exports[key]?.['require'];
-    if (typeof entry === 'object') {
-      entry = entry['require'] ?? entry['default'];
-    }
-  }
-
-  return entry;
-}
-
+/**
+ * @deprecated Use {@link fromPackageJson} instead, which offers a fluent,
+ * composable API for skipping, overriding and patching shared externals:
+ *
+ * ```ts
+ * // before
+ * shareAll(config, { skipList, overrides });
+ * // after
+ * fromPackageJson(config).skip(skipList).override(overrides).get();
+ * ```
+ *
+ * Note: `fromPackageJson().skip()` *extends* the default skip list, whereas
+ * `shareAll`'s `skipList` option *replaces* it.
+ */
 export function shareAll(
   config: ShareAllExternalsOptions,
   opts: {
@@ -371,17 +90,11 @@ export function shareAllCore(
     skipList?: SkipList;
     projectPath?: string;
     overrides?: ShareExternalsOptions;
+    patchList?: Record<string, Partial<ExternalConfig>>;
   } = {},
   repo: PackageJsonRepository = sharedPackageJsonRepository
 ): ResolvedSharedExternalsConfig {
-  // let workspacePath: string | undefined = undefined;
   const projectPath = inferProjectPath(opts.projectPath);
-
-  // workspacePath = getConfigContext().workspaceRoot ?? '';
-
-  // if (!workspacePath) {
-  //   workspacePath = projectPath;
-  // }
 
   const versionMaps = getVersionMaps(projectPath, projectPath, repo);
   const sharedExternals: ShareExternalsOptions = {};
@@ -405,29 +118,47 @@ export function shareAllCore(
     }
   }
 
+  const finalExternalList = applyPatchList(sharedExternals, opts.patchList, opts.overrides);
+
   return {
-    ...shareCore(io, sharedExternals, opts.projectPath, skipList, repo),
+    ...shareCore(io, finalExternalList, opts.projectPath, skipList, repo),
     ...(!opts.overrides ? {} : shareCore(io, opts.overrides, opts.projectPath, skipList, repo)),
   };
 }
 
-function inferProjectPath(projectPath: string | undefined) {
-  if (!projectPath && getConfigContext().packageJson) {
-    projectPath = path.dirname(getConfigContext().packageJson || '');
+/**
+ * Merges `patchList` overrides onto the shared externals. Patches only affect
+ * externals that are actually being shared: a patch for an external that isn't
+ * in the list (unknown dependency, skipped, or shadowed by `overrides`) is
+ * ignored with a warning.
+ */
+function applyPatchList(
+  sharedExternals: ShareExternalsOptions,
+  patchList: Record<string, Partial<ExternalConfig>> | undefined,
+  overrides: ShareExternalsOptions | undefined
+): ShareExternalsOptions {
+  if (!patchList) {
+    return sharedExternals;
   }
 
-  if (!projectPath && getConfigContext().workspaceRoot) {
-    projectPath = getConfigContext().workspaceRoot || '';
+  const result = { ...sharedExternals };
+
+  for (const [external, cfg] of Object.entries(patchList)) {
+    if (!result[external]) {
+      const shadowedByOverride =
+        !!overrides && Object.keys(overrides).some(o => external.startsWith(o));
+      logger.warn(
+        shadowedByOverride
+          ? `Ignoring patch for '${external}': it is already configured via 'overrides' ('patch' and 'overrides' are mutually exclusive per external).`
+          : `Ignoring patch for '${external}': it is not a shared external (unknown dependency or skipped).`
+      );
+      continue;
+    }
+
+    result[external] = { ...result[external], ...cfg };
   }
 
-  if (!projectPath) {
-    projectPath = cwd();
-  }
-  return projectPath;
-}
-
-export function setInferVersion(infer: boolean): void {
-  inferVersion = infer;
+  return result;
 }
 
 export function share(
@@ -452,17 +183,16 @@ export function shareCore(
 
   const shareObjects = { ...configuredShareObjects };
 
-  const result: any = {};
-  let includeSecondaries;
+  const result: SharedExternalsConfig = {};
 
   for (const key in shareObjects) {
-    includeSecondaries = false;
-    const shareObject = (shareObjects as any)[key];
+    let includeSecondaries: IncludeSecondariesOptions = false;
+    const shareObject = shareObjects[key]!;
 
     if (
       shareObject.requiredVersion === 'auto' ||
-      (inferVersion && typeof shareObject.requiredVersion === 'undefined') ||
-      shareObject.requiredVersion?.length < 1
+      (isInferVersion() && typeof shareObject.requiredVersion === 'undefined') ||
+      (shareObject.requiredVersion?.length ?? 1) < 1
     ) {
       const version = lookupVersion(key, projectPath, repo);
 
@@ -477,7 +207,9 @@ export function shareCore(
     if (shareObject.includeSecondaries) {
       includeSecondaries = shareObject.includeSecondaries;
       delete shareObject.includeSecondaries;
-      if (includeSecondaries?.keepAll) shareObject.includeSecondaries = true;
+      if (typeof includeSecondaries === 'object' && includeSecondaries.keepAll) {
+        shareObject.includeSecondaries = true;
+      }
     }
 
     result[key] = shareObject;
@@ -506,14 +238,5 @@ export function shareCore(
     }
   }
 
-  return result;
-}
-
-function addSecondaries(
-  secondaries: Record<string, ExternalConfig>,
-  result: Record<string, ExternalConfig>
-) {
-  for (const key in secondaries) {
-    result[key] = secondaries[key]!;
-  }
+  return result as ResolvedSharedExternalsConfig;
 }
